rename all files of musicalbum
