+++ classes/musicalbum_helper.rb
@@ -1,7 +1,7 @@
-require_relative 'music_album'
-require_relative '../modules/add_music_album'
-require_relative '../modules/list_music_albums'
-require_relative '../modules/preserve_music_album'
+require_relative 'musicalbum'
+require_relative '../modules/add_musicalbum'
+require_relative '../modules/list_musicalbums'
+require_relative '../modules/preserve_musicalbum'
 require_relative 'genre'
 require_relative '../modules/add_genre'
 require_relative '../modules/list_genres'
